+++ posthog-react-native/CHANGELOG.md
@@ -1,4 +1,9 @@
-# Next
+# Next (v4.0.0)
+
+1. Remove deprecated `personProperties` and `groupProperties`
+2. Rename `captureNativeAppLifecycleEvents` option to `captureAppLifecycleEvents`
+   1. `captureAppLifecycleEvents` from `autocapture` is removed and replaced by `captureAppLifecycleEvents` from options
+3. Removed `version` and `build` from Lifecycle events in favor of `$app_version` and `$app_build`
 
 # 3.16.1 – 2025-05-28